Tarea 2 corregida
se hicieron cambios totales en el hardware como cables y pines. Falta poner la FSM y el Magic sigue teniendo su warning.

+++ PeripheralsDrivers/Src/systick_driver_hal.c
@@ -61,7 +61,7 @@ void systick_Delay_ms(uint32_t wait_time_ms){
 
 	ticks_counting = getTicks_ms();
 
-	while(ticks_counting < (ticks_start + (uint64_t) wait_time_ms)){
+	while(ticks_counting < (ticks_start + (uint32_t) wait_time_ms)){
 		//actualizar el valor
 			ticks_counting = getTicks_ms();
